Add: WithExternalCreatedBy mixin;
