+++ src/ggrc/models/control.py
@@ -14,6 +14,7 @@ from ggrc.models.mixins.with_similarity_score import WithSimilarityScore
 from ggrc.models.object_document import PublicDocumentable
 from ggrc.models.mixins import base, categorizable
 from ggrc.models.mixins import synchronizable
+from ggrc.models.mixins import with_external_created_by
 from ggrc.models import mixins, utils
 from ggrc.models.mixins.with_last_assessment_date import WithLastAssessmentDate
 from ggrc.models.deferred import deferred
@@ -25,7 +26,8 @@ from ggrc.models import reflection
 from ggrc.models.exceptions import ValidationError
 
 
-class Control(synchronizable.Synchronizable,
+class Control(with_external_created_by.WithExternalCreatedBy,
+              synchronizable.Synchronizable,
               categorizable.Categorizable,
               WithLastAssessmentDate,
               synchronizable.RoleableSynchronizable,
@@ -61,14 +63,6 @@ class Control(synchronizable.Synchronizable,
 
   # GGRCQ attributes
   due_date = db.Column(db.Date, nullable=True)
-  created_by_id = db.Column(db.Integer, nullable=False)
-
-  # pylint: disable=no-self-argument
-  @declared_attr
-  def created_by(cls):
-    """Relationship to user referenced by created_by_id."""
-    return utils.person_relationship(cls.__name__, "created_by_id")
-
   last_submitted_at = db.Column(db.DateTime, nullable=True)
   last_submitted_by_id = db.Column(db.Integer, nullable=True)
 
@@ -90,7 +84,6 @@ class Control(synchronizable.Synchronizable,
   _title_uniqueness = False
 
   _custom_publish = {
-      'created_by': ggrc_utils.created_by_stub,
       'last_submitted_by': ggrc_utils.last_submitted_by_stub,
       'last_verified_by': ggrc_utils.last_verified_by_stub,
   }
@@ -109,8 +102,6 @@ class Control(synchronizable.Synchronizable,
       'review_status',
       'review_status_display_name',
       'due_date',
-      reflection.ExternalUserAttribute('created_by',
-                                       force_create=True),
       'last_submitted_at',
       reflection.ExternalUserAttribute('last_submitted_by',
                                        force_create=True),
