+++ src/ggrc/models/mixins/__init__.py
@@ -36,6 +36,7 @@ from ggrc.models.deferred import deferred
 from ggrc.models.mixins.base import Base
 from ggrc.models.mixins.customattributable import CustomAttributable
 from ggrc.models.mixins.notifiable import Notifiable
+from ggrc.models.mixins.with_external_created_by import WithExternalCreatedBy
 from ggrc.models.utils import validate_option
 from ggrc.fulltext import attributes
 
@@ -652,7 +653,7 @@ class BusinessObject(Stateful, Noted, Described, Titled, Slugged):
   }
 
 
-class ScopeObject(BusinessObject):
+class ScopeObject(WithExternalCreatedBy, BusinessObject):
   """Mixin that re-name status attribute"""
 
   _fulltext_attrs = [
